+++ litho-widget/src/main/java/com/facebook/litho/widget/ComponentRenderInfo.java
@@ -23,6 +23,7 @@ import com.facebook.litho.ComponentContext;
 import com.facebook.litho.ComponentsLogger;
 import com.facebook.litho.EventHandler;
 import com.facebook.litho.RenderCompleteEvent;
+import com.facebook.litho.RenderResult;
 
 /** {@link RenderInfo} that can render components. */
 public class ComponentRenderInfo extends BaseRenderInfo {
@@ -131,8 +132,8 @@ public class ComponentRenderInfo extends BaseRenderInfo {
   private static class EmptyComponent extends Component {
 
     @Override
-    protected Component onCreateLayout(ComponentContext c) {
-      return Column.create(c).build();
+    protected RenderResult render(ComponentContext c) {
+      return new RenderResult(Column.create(c).build());
     }
 
     @Override